verifycname/requirements.txt
gradio==4.41.0
dnspython==2.6.1